backend/requirements.txt
fastapi
httpx
psycopg[binary]
python-dotenv
uvicorn[standard]
